HF_Space_Backend/requirements.txt
fastapi[all]
python-dotenv
torch
transformers
coqui-tts
huggingface_hub
requests
python-multipart
aiofiles
uvicorn
regex